@@ -10,7 +10,7 @@ buy_omen = "scripts.bet_omen:buy"
 sell_omen = "scripts.bet_omen:sell"
 
 [tool.poetry.dependencies]
-python = ">=3.9,<3.12"
+python = ">=3.10,<3.12"
 typer = "^0.9.0"
 types-requests = "^2.31.0.20240106"
 google-cloud-functions = "^1.16.0"
@@ -22,6 +22,7 @@ web3 = "^6.15.1"
 eth-typing = "^4.0.0"
 pydantic-settings = "^2.1.0"
 numpy = "^1.26.4"
+streamlit = "^1.31.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "*"
